export select pipette part as PipetteConfiguration component
