+++ protocol-designer/src/organisms/EditInstrumentsModal/index.tsx
@@ -2,68 +2,33 @@ import { useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
 import { createPortal } from 'react-dom'
-import styled, { css } from 'styled-components'
 
 import {
-  ALIGN_CENTER,
-  ALIGN_STRETCH,
-  Box,
-  Btn,
-  Checkbox,
-  COLORS,
-  CURSOR_POINTER,
-  DIRECTION_COLUMN,
-  DISPLAY_FLEX,
-  DISPLAY_INLINE_BLOCK,
-  FLEX_MAX_CONTENT,
   Flex,
   JUSTIFY_END,
   Modal,
-  OVERFLOW_AUTO,
   PrimaryButton,
-  PRODUCT,
-  RadioButton,
   SecondaryButton,
   SPACING,
-  StyledText,
-  TYPOGRAPHY,
-  WRAP,
 } from '@opentrons/components'
-import {
-  FLEX_ROBOT_TYPE,
-  getAllPipetteNames,
-  OT2_ROBOT_TYPE,
-} from '@opentrons/shared-data'
 
-import { getAllowAllTipracks } from '../../feature-flags/selectors'
 import {
   getAdditionalEquipment,
   getInitialDeckSetup,
   getPipetteEntities,
 } from '../../step-forms/selectors'
-import { getHas96Channel, removeOpentronsPhrases } from '../../utils'
-
+import { getHas96Channel } from '../../utils'
 import { getRobotType } from '../../file-data/selectors'
-import {
-  PIPETTE_GENS,
-  PIPETTE_TYPES,
-  PIPETTE_VOLUMES,
-} from '../../pages/CreateNewProtocolWizard/constants'
-import { getLabwareDefsByURI } from '../../labware-defs/selectors'
-import { setFeatureFlags } from '../../feature-flags/actions'
-import { createCustomTiprackDef } from '../../labware-defs/actions'
 import { selectors as stepFormSelectors } from '../../step-forms'
 import { getMainPagePortalEl } from '../Portal'
-import { getShouldShowPipetteType, getTiprackOptions } from './utils'
 import { editPipettes } from './editPipettes'
 import { HandleEnter } from '../../atoms/HandleEnter'
 import { PipetteOverview } from './PipetteOverview'
+import { PipetteConfiguration } from './PipetteConfiguration'
 
 import type { PipetteMount, PipetteName } from '@opentrons/shared-data'
 import type {
   Gen,
-  PipetteInfoByGen,
-  PipetteInfoByType,
   PipetteType,
 } from '../../pages/CreateNewProtocolWizard/types'
 import type { ThunkDispatch } from '../../types'
@@ -88,14 +53,11 @@ export function EditInstrumentsModal(
   const [pipetteGen, setPipetteGen] = useState<Gen | 'flex'>('flex')
   const [pipetteVolume, setPipetteVolume] = useState<string | null>(null)
   const [selectedTips, setSelectedTips] = useState<string[]>([])
-  const allowAllTipracks = useSelector(getAllowAllTipracks)
   const robotType = useSelector(getRobotType)
   const orderedStepIds = useSelector(stepFormSelectors.getOrderedStepIds)
   const initialDeckSetup = useSelector(getInitialDeckSetup)
   const additionalEquipment = useSelector(getAdditionalEquipment)
   const pipetteEntities = useSelector(getPipetteEntities)
-  const allLabware = useSelector(getLabwareDefsByURI)
-  const allPipetteOptions = getAllPipetteNames('maxVolume', 'channels')
   const { pipettes, labware } = initialDeckSetup
   const pipettesOnDeck = Object.values(pipettes)
   const has96Channel = getHas96Channel(pipetteEntities)
@@ -104,7 +66,7 @@ export function EditInstrumentsModal(
   const gripper = Object.values(additionalEquipment).find(
     ae => ae.name === 'gripper'
   )
-  const selectedPip =
+  const selectedPipette =
     pipetteType === '96' || pipetteGen === 'GEN1'
       ? `${pipetteVolume}_${pipetteType}`
       : `${pipetteVolume}_${pipetteType}_${pipetteGen.toLowerCase()}`
@@ -126,7 +88,7 @@ export function EditInstrumentsModal(
         orderedStepIds,
         dispatch,
         mount,
-        selectedPip as PipetteName,
+        selectedPipette as PipetteName,
         selectedTips,
         leftPipette,
         rightPipette
@@ -200,223 +162,25 @@ export function EditInstrumentsModal(
             gripper={gripper}
           />
         ) : (
-          <Flex
-            flexDirection="column"
-            overflowY={OVERFLOW_AUTO}
-            gridGap={SPACING.spacing24}
-          >
-            <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing8}>
-              <StyledText desktopStyle="bodyLargeSemiBold">
-                {t('pipette_type')}
-              </StyledText>
-              <Flex gridGap={SPACING.spacing4}>
-                {PIPETTE_TYPES[robotType].map(type => {
-                  return getShouldShowPipetteType(
-                    type.value as PipetteType,
-                    has96Channel,
-                    leftPipette,
-                    rightPipette,
-                    mount
-                  ) ? (
-                    <RadioButton
-                      key={`${type.label}_${type.value}`}
-                      onChange={() => {
-                        setPipetteType(type.value)
-                        setPipetteGen('flex')
-                        setPipetteVolume(null)
-                        setSelectedTips([])
-                      }}
-                      buttonLabel={t(`shared:${type.label}`)}
-                      buttonValue="single"
-                      isSelected={pipetteType === type.value}
-                    />
-                  ) : null
-                })}
-              </Flex>
-            </Flex>
-            {pipetteType != null && robotType === OT2_ROBOT_TYPE ? (
-              <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing8}>
-                <StyledText desktopStyle="bodyLargeSemiBold">
-                  {t('pipette_gen')}
-                </StyledText>
-                <Flex gridGap={SPACING.spacing4}>
-                  {PIPETTE_GENS.map(gen => (
-                    <RadioButton
-                      key={gen}
-                      onChange={() => {
-                        setPipetteGen(gen)
-                        setPipetteVolume(null)
-                        setSelectedTips([])
-                      }}
-                      buttonLabel={gen}
-                      buttonValue={gen}
-                      isSelected={pipetteGen === gen}
-                    />
-                  ))}
-                </Flex>
-              </Flex>
-            ) : null}
-            {(pipetteType != null && robotType === FLEX_ROBOT_TYPE) ||
-            (pipetteGen !== 'flex' &&
-              pipetteType != null &&
-              robotType === OT2_ROBOT_TYPE) ? (
-              <Flex
-                flexDirection={DIRECTION_COLUMN}
-                gridGap={SPACING.spacing8}
-                id="volume"
-              >
-                <StyledText desktopStyle="bodyLargeSemiBold">
-                  {t('pipette_vol')}
-                </StyledText>
-                <Flex gridGap={SPACING.spacing4}>
-                  {PIPETTE_VOLUMES[robotType]?.map(volume => {
-                    if (robotType === FLEX_ROBOT_TYPE && pipetteType != null) {
-                      const flexVolume = volume as PipetteInfoByType
-                      const flexPipetteInfo = flexVolume[pipetteType]
-
-                      return flexPipetteInfo?.map(type => (
-                        <RadioButton
-                          key={`${type.value}_${pipetteType}`}
-                          onChange={() => {
-                            setPipetteVolume(type.value)
-                            setSelectedTips([])
-                          }}
-                          buttonLabel={t('vol_label', { volume: type.label })}
-                          buttonValue={type.value}
-                          isSelected={pipetteVolume === type.value}
-                        />
-                      ))
-                    } else {
-                      const ot2Volume = volume as PipetteInfoByGen
-                      const gen = pipetteGen as Gen
-
-                      return ot2Volume[gen].map(info => {
-                        return info[pipetteType]?.map(type => (
-                          <RadioButton
-                            key={`${type.value}_${pipetteGen}_${pipetteType}`}
-                            onChange={() => {
-                              setPipetteVolume(type.value)
-                            }}
-                            buttonLabel={t('vol_label', {
-                              volume: type.label,
-                            })}
-                            buttonValue={type.value}
-                            isSelected={pipetteVolume === type.value}
-                          />
-                        ))
-                      })
-                    }
-                  })}
-                </Flex>
-              </Flex>
-            ) : null}
-            {allPipetteOptions.includes(selectedPip as PipetteName)
-              ? (() => {
-                  const tiprackOptions = getTiprackOptions({
-                    allLabware,
-                    allowAllTipracks,
-                    selectedPipetteName: selectedPip,
-                  })
-                  return (
-                    <Flex
-                      flexDirection={DIRECTION_COLUMN}
-                      gridGap={SPACING.spacing8}
-                    >
-                      <StyledText desktopStyle="bodyLargeSemiBold">
-                        {t('pipette_tips')}
-                      </StyledText>
-                      <Box
-                        css={css`
-                          gap: ${SPACING.spacing4};
-                          display: ${DISPLAY_FLEX};
-                          flex-wrap: ${WRAP};
-                          align-items: ${ALIGN_CENTER};
-                          align-content: ${ALIGN_CENTER};
-                          align-self: ${ALIGN_STRETCH};
-                        `}
-                      >
-                        {tiprackOptions.map(option => (
-                          <Checkbox
-                            key={option.value}
-                            disabled={
-                              selectedTips.length === 3 &&
-                              !selectedTips.includes(option.value)
-                            }
-                            isChecked={selectedTips.includes(option.value)}
-                            labelText={removeOpentronsPhrases(option.name)}
-                            onClick={() => {
-                              const updatedTips = selectedTips.includes(
-                                option.value
-                              )
-                                ? selectedTips.filter(v => v !== option.value)
-                                : [...selectedTips, option.value]
-                              setSelectedTips(updatedTips)
-                            }}
-                          />
-                        ))}
-                        <Flex
-                          gridGap={SPACING.spacing8}
-                          padding={SPACING.spacing4}
-                          width={FLEX_MAX_CONTENT}
-                        >
-                          <StyledLabel>
-                            <StyledText desktopStyle="bodyDefaultRegular">
-                              {t('add_custom_tips')}
-                            </StyledText>
-                            <input
-                              data-testid="SelectPipettes_customTipInput"
-                              type="file"
-                              onChange={e =>
-                                dispatch(createCustomTiprackDef(e))
-                              }
-                            />
-                          </StyledLabel>
-                          {pipetteVolume === 'p1000' &&
-                          robotType === FLEX_ROBOT_TYPE ? null : (
-                            <Btn
-                              onClick={() => {
-                                dispatch(
-                                  setFeatureFlags({
-                                    OT_PD_ALLOW_ALL_TIPRACKS: !allowAllTipracks,
-                                  })
-                                )
-                              }}
-                              textDecoration={
-                                TYPOGRAPHY.textDecorationUnderline
-                              }
-                            >
-                              <StyledLabel>
-                                <StyledText desktopStyle="bodyDefaultRegular">
-                                  {allowAllTipracks
-                                    ? t('show_default_tips')
-                                    : t('show_all_tips')}
-                                </StyledText>
-                              </StyledLabel>{' '}
-                            </Btn>
-                          )}
-                        </Flex>
-                      </Box>
-                    </Flex>
-                  )
-                })()
-              : null}
-          </Flex>
+          <PipetteConfiguration
+            has96Channel={has96Channel}
+            robotType={robotType}
+            pipetteType={pipetteType}
+            pipetteGen={pipetteGen}
+            mount={mount}
+            pipetteVolume={pipetteVolume}
+            selectedTips={selectedTips}
+            selectedPipette={selectedPipette}
+            setPipetteGen={setPipetteGen}
+            setPipetteType={setPipetteType}
+            setPipetteVolume={setPipetteVolume}
+            setSelectedTips={setSelectedTips}
+            leftPipette={leftPipette}
+            rightPipette={rightPipette}
+          />
         )}
       </Modal>
     </HandleEnter>,
     getMainPagePortalEl()
   )
 }
-
-const StyledLabel = styled.label`
-  text-decoration: ${TYPOGRAPHY.textDecorationUnderline};
-  font-size: ${PRODUCT.TYPOGRAPHY.fontSizeBodyDefaultSemiBold};
-  display: ${DISPLAY_INLINE_BLOCK};
-  cursor: ${CURSOR_POINTER};
-  input[type='file'] {
-    display: none;
-  }
-  &:hover {
-    color: ${COLORS.blue50};
-  }
-`
